requirements.txt
black
jinja2
pytest
pytest-django
rs-parsepatch
ruff
uwsgi
networkx
kombu
python-jose
